+++ execution/__init__.py
@@ -1,0 +1,15 @@
+from .execution_engine import ExecutionEngine
+from .execution_context import ExecutionContext
+from .dependency_manager import DependencyManager, DependencyGraph
+from .task_scheduler import TaskScheduler, AtomicStage
+from .stage_executor import StageExecutor
+
+__all__ = [
+    "ExecutionEngine",
+    "ExecutionContext",
+    "DependencyManager",
+    "DependencyGraph",
+    "TaskScheduler",
+    "AtomicStage",
+    "StageExecutor",
+]